refactor(plugin-data-source-manager): split v2 collection templates
