+++ packages/plugins/@nocobase/plugin-collection-sql/client-v2.js
@@ -1,0 +1,3 @@
+'use strict';
+
+module.exports = require('./dist/client-v2/index.js');
