+++ packages/plugins/@nocobase/plugin-collection-sql/package.json
@@ -14,6 +14,9 @@
   "license": "Apache-2.0",
   "peerDependencies": {
     "@nocobase/client": "2.x",
+    "@nocobase/client-v2": "2.x",
+    "@nocobase/flow-engine": "2.x",
+    "@nocobase/plugin-data-source-manager": "2.x",
     "@nocobase/server": "2.x",
     "@nocobase/test": "2.x"
   },
