+++ packages/plugins/@nocobase/plugin-collection-sql/src/client-v2/index.tsx
@@ -1,0 +1,12 @@
+/**
+ * This file is part of the NocoBase (R) project.
+ * Copyright (c) 2020-2024 NocoBase Co., Ltd.
+ * Authors: NocoBase Team.
+ *
+ * This project is dual-licensed under AGPL-3.0 and NocoBase Commercial License.
+ * For more information, please refer to: https://www.nocobase.com/agreement.
+ */
+
+export { default } from './plugin';
+export * from './plugin';
+export * from './SqlCollectionConfigure';
