+++ packages/plugins/@nocobase/plugin-collection-tree/client-v2.d.ts
@@ -1,0 +1,2 @@
+export * from './dist/client-v2';
+export { default } from './dist/client-v2';
